+++ trunk/src/weapon/snipe_rifle.cpp
@@ -122,7 +122,7 @@ void SnipeRifle::ComputeCrossPoint(bool force = false)
 {
   // Did the current character is moving ?
   Point2i pos = GetGunHolePosition();
-  double angle = ActiveCharacter().GetFiringAngle();
+  Double angle = ActiveCharacter().GetFiringAngle();
   if ( !force && last_rifle_pos == pos && last_angle == angle )
     return;
   else
@@ -132,7 +132,7 @@ void SnipeRifle::ComputeCrossPoint(bool force = false)
   }
 
   // Equation of movement : y = ax + b
-  double a, b;
+  Double a, b;
   a = sin(angle) / cos(angle);
   b = pos.y - ( a * pos.x );
   Point2i delta_pos, size, start_point;
@@ -183,7 +183,7 @@ void SnipeRifle::DrawBeam()
 {
   Point2i pos1 = laser_beam_start - Camera::GetInstance()->GetPosition();
   Point2i pos2 = targeted_point - Camera::GetInstance()->GetPosition();
-  float dst = laser_beam_start.Distance(targeted_point);
+  Double dst = laser_beam_start.Distance(targeted_point);
 
   GetMainWindow().
     AAFadingLineColor(pos1.x, pos2.x, pos1.y, pos2.y, laser_beam_color, Color(255, 0, 0, 0));
@@ -192,9 +192,9 @@ void SnipeRifle::DrawBeam()
 
   // Set area of the screen to be redrawn:
   // Splited into little rectangles to avoid too large area of redraw
-  float redraw_size = 20.0;
-  Point2f pos = Point2f((float)laser_beam_start.x, (float)laser_beam_start.y);
-  Point2f delta = ( Point2f((float)targeted_point.x, (float)targeted_point.y) - pos ) * redraw_size / dst;
+  Double redraw_size = 20.0;
+  Point2f pos = Point2f((Double)laser_beam_start.x, (Double)laser_beam_start.y);
+  Point2f delta = ( Point2f((Double)targeted_point.x, (Double)targeted_point.y) - pos ) * redraw_size / dst;
   Point2i delta_i((int)delta.x, (int)delta.y);
 
   if(delta_i.x < 0) delta_i.x = - delta_i.x; // the Map::ToRedraw method doesn't support negative rectangles
@@ -203,9 +203,9 @@ void SnipeRifle::DrawBeam()
   delta_i.y += 6;
 
   int i = 0;
-  while( (float)i * redraw_size < dst )
+  while( (Double)i * redraw_size < dst )
   {
-    // float to int conversion...
+    // Double to int conversion...
     Point2i pos_i((int)pos.x, (int)pos.y);
     if(delta.x < 0.0)
     {
